move new types to types.rs
